Finishing up the main idea
Everything works, and the file structure was tweaked to work with jar
distribution.
For now, I consider a little bow tied on top of this project. It does
what it set out to do, and i will be able to gather even more words in
the list, if i ever need such a thing for future use.