Fixed a method, added an Evaluator class for score and minimax.
Fixed a problem when a move is played. a new array was supposed to be operated on, but instead it worked on the original one. This wasn't a problem while it was just the one board that i was working with, but when more boards were created as children nodes, it just played all the possible moves on the original board, filling it up.

The Evaluator class is my answer to the question of where to put the minimax algorithm in code. I also moved the scoring method for individual boards to that class, since it fit logically. Currently it can evaluate a single board, and all that is really left is the minimax implementation. which sounds like a lot of work honestly.