Implemented (for the third time) the board score algorithm, and other other things
Changed the minimax code by one line, because i switched up the order by accident.
Implemented a very rough and ugly threading process. As it stands, everything is threaded except the code that blocks the other player from winning.

A question that just occured to me is, am i allowed to make board decisions outside of my minimax algorithm? because the blocking move excecutes before all of that.